requirements.txt
django
pillow
mysqlclient
djangorestframework
markdown
django-filter
django-crispy-forms
django-reversion
django-formtools
six
coreapi
django-cors-headers
drf-extensions
django-redis
pycryptodome
raven
django-import-export
itsdangerous
cryptography
msgpack
celery
eventlet
django-celery
django-jsonfield
django-mptt
pymysql
sshtunnel
Django-extensions
pandas
pysimdjson
pycryptodomex
django-mptt
numba
pysimdjson
redis==2.10.6
billiard==3.6